// src/types/dataTypes/marketsData.ts
export type MarketsDataType = {
  symbol: string;
  lastPrice: string;
  _24hrHighPrice: string;
  _24hrLowPrice: string;
  _24hrVolume: string;
  _24hrCount: string;
  indexPrice: string;
  lastFundingRate: string;
  nextFundingTime: number;
  avgFundingRate: string;
  _24hrPriceChange: string;
  _24hrPriceChangePercent: string;
  marketPrice: string;
  marketPriceDirection: -1 | 1;
};